+++ pytorch_lightning/callbacks/gpu_stats_monitor.py
@@ -1,6 +1,20 @@
+# Copyright The PyTorch Lightning team.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 """
 GPU Stats Monitor
-====================
+=================
 
 Monitor and logs GPU stats during training.
 
@@ -12,26 +26,26 @@ import subprocess
 import time
 
 from pytorch_lightning.callbacks.base import Callback
-from pytorch_lightning.utilities import rank_zero_only, rank_zero_warn
+from pytorch_lightning.utilities import rank_zero_only
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from pytorch_lightning.utilities.parsing import AttributeDict
 
 
 class GPUStatsMonitor(Callback):
     r"""
-    Automatically monitor and logs GPU stats during training stage. ``GPUStatsMonitor``
+    Automatically monitors and logs GPU stats during training stage. ``GPUStatsMonitor``
     is a callback and in order to use it you need to assign a logger in the ``Trainer``.
 
     Args:
-        memory_utilization: Set to ``True`` to log used, free and percentage of memory
+        memory_utilization: Set to ``True`` to monitor used, free and percentage of memory
             utilization at the start and end of each step. Default: ``True``.
-        gpu_utilization: Set to ``True`` to log percentage of GPU utilization
+        gpu_utilization: Set to ``True`` to monitor percentage of GPU utilization
             at the start and end of each step. Default: ``True``.
-        intra_step_time: Set to ``True`` to log the time of each step. Default: ``False``.
-        inter_step_time: Set to ``True`` to log the time between the end of one step
+        intra_step_time: Set to ``True`` to monitor the time of each step. Default: ``False``.
+        inter_step_time: Set to ``True`` to monitor the time between the end of one step
             and the start of the next step. Default: ``False``.
-        fan_speed: Set to ``True`` to log percentage of fan speed. Default: ``False``.
-        temperature: Set to ``True`` to log the memory and gpu temperature in degree Celsius.
+        fan_speed: Set to ``True`` to monitor percentage of fan speed. Default: ``False``.
+        temperature: Set to ``True`` to monitor the memory and gpu temperature in degree Celsius.
             Default: ``False``.
 
     Example::
@@ -69,7 +83,7 @@ class GPUStatsMonitor(Callback):
     ):
         super().__init__()
 
-        if shutil.which("nvidia-smi") is None:
+        if shutil.which('nvidia-smi') is None:
             raise MisconfigurationException(
                 'Cannot use GPUStatsMonitor callback because NVIDIA driver is not installed.'
             )
@@ -90,78 +104,90 @@ class GPUStatsMonitor(Callback):
             )
 
         if not trainer.on_gpu:
-            rank_zero_warn(
-                'You are using GPUStatsMonitor but are not running on GPU.'
-                ' Logged utilization will be independent from your model.', RuntimeWarning
+            raise MisconfigurationException(
+                'You are using GPUStatsMonitor but are not running on GPU'
+                f' since gpus attribute in Trainer is set to {trainer.gpus}.'
             )
 
+        self._gpu_ids = ','.join(map(str, trainer.data_parallel_device_ids))
+
     def on_train_epoch_start(self, trainer, pl_module):
-        self.snap_intra_step_time = None
-        self.snap_inter_step_time = None
+        self._snap_intra_step_time = None
+        self._snap_inter_step_time = None
 
     @rank_zero_only
     def on_train_batch_start(self, trainer, pl_module, batch, batch_idx, dataloader_idx):
-        if self._log_stats.gpu_utilization:
-            self._log_usage(trainer)
+        gpu_stat_keys = self._get_gpu_stat_keys()
+        gpu_stats = self._get_gpu_stats(gpu_stat_keys)
 
-        if self._log_stats.memory_utilization:
-            self._log_memory(trainer)
-
-        if self._log_stats.inter_step_time and self.snap_inter_step_time:
+        if self._log_stats.inter_step_time and self._snap_inter_step_time:
             # First log at beginning of second step
-            trainer.logger.log_metrics(
-                {'batch_time/inter_step (ms)': (time.time() - self.snap_inter_step_time) * 1000},
-                step=trainer.global_step
-            )
+            gpu_stats['batch_time/inter_step (ms)'] = (time.time() - self._snap_inter_step_time) * 1000
 
         if self._log_stats.intra_step_time:
-            self.snap_intra_step_time = time.time()
+            self._snap_intra_step_time = time.time()
+
+        trainer.logger.log_metrics(gpu_stats, step=trainer.global_step)
 
     @rank_zero_only
     def on_train_batch_end(self, trainer, pl_module, batch, batch_idx, dataloader_idx):
-        if self._log_stats.gpu_utilization:
-            self._log_usage(trainer)
-
-        if self._log_stats.memory_utilization:
-            self._log_memory(trainer)
-
-        if self._log_stats.fan_speed:
-            trainer.logger.log_metrics(self._get_gpu_stat("fan.speed", "%"), step=trainer.global_step)
-
-        if self._log_stats.temperature:
-            trainer.logger.log_metrics(self._get_gpu_stat("temperature.gpu", "degrees C"), step=trainer.global_step)
-            trainer.logger.log_metrics(self._get_gpu_stat("temperature.memory", "degrees C"), step=trainer.global_step)
-
+        gpu_stat_keys = self._get_gpu_stat_keys() + self._get_gpu_device_stat_keys()
+        gpu_stats = self._get_gpu_stats(gpu_stat_keys)
+        
         if self._log_stats.inter_step_time:
-            self.snap_inter_step_time = time.time()
+            self._snap_inter_step_time = time.time()
 
-        if self._log_stats.intra_step_time and self.snap_intra_step_time:
-            trainer.logger.log_metrics(
-                {'batch_time/intra_step (ms)': (time.time() - self.snap_intra_step_time) * 1000},
-                step=trainer.global_step
-            )
+        if self._log_stats.intra_step_time and self._snap_intra_step_time:
+            gpu_stats['batch_time/intra_step (ms)'] = (time.time() - self._snap_intra_step_time) * 1000
 
-    @staticmethod
-    def _get_gpu_stat(pitem: str, unit: str):
+        trainer.logger.log_metrics(gpu_stats, step=trainer.global_step)
+
+    def _get_gpu_stats(self, gpu_stat_keys):
+        gpu_query = ','.join([m[0] for m in gpu_stat_keys])
+        format = 'csv,nounits,noheader'
+
         result = subprocess.run(
-            [shutil.which("nvidia-smi"), f"--query-gpu={pitem}", "--format=csv,nounits,noheader"],
+            [shutil.which('nvidia-smi'), f'--query-gpu={gpu_query}', f'--format={format}', f'--id={self._gpu_ids}'],
             encoding="utf-8",
             stdout=subprocess.PIPE,
             stderr=subprocess.PIPE,  # for backward compatibility with python version 3.6
             check=True
         )
 
-        try:
-            gpu_usage = [float(x) for x in result.stdout.strip().split(os.linesep)]
-        except ValueError:
-            gpu_usage = [0]
+        def _to_float(x):
+            try:
+                return float(x)
+            except ValueError:
+                return 0.
 
-        return {f"gpu_{pitem}/gpu_id_{index} ({unit})": usage for index, usage in enumerate(gpu_usage)}
+        stats = result.stdout.strip().split(os.linesep)
+        stats = [list(map(_to_float, x.split(', '))) for x in stats]
 
-    def _log_usage(self, trainer):
-        trainer.logger.log_metrics(self._get_gpu_stat("utilization.gpu", "%"), step=trainer.global_step)
+        logs = {}
+        for i, gpu_id in enumerate(self._gpu_ids.split(',')):
+            gpu_stat_keys = [f'gpu_id: {gpu_id}/{x} ({unit})' for x, unit in gpu_stat_keys]
+            logs.update(dict(zip(gpu_stat_keys, stats[i])))
 
-    def _log_memory(self, trainer):
-        trainer.logger.log_metrics(self._get_gpu_stat("memory.used", "MB"), step=trainer.global_step)
-        trainer.logger.log_metrics(self._get_gpu_stat("memory.free", "MB"), step=trainer.global_step)
-        trainer.logger.log_metrics(self._get_gpu_stat("utilization.memory", "%"), step=trainer.global_step)
+        return logs
+
+    def _get_gpu_stat_keys(self):
+        stat_keys = []
+
+        if self._log_stats.gpu_utilization:
+            stat_keys.append(('utilization.gpu', '%'))
+
+        if self._log_stats.memory_utilization:
+            stat_keys.extend([('memory.used', 'MB'), ('memory.free', 'MB'), ('utilization.memory', '%')])
+
+        return stat_keys
+
+    def _get_gpu_device_stat_keys(self):
+        stat_keys = []
+
+        if self._log_stats.fan_speed:
+            stat_keys.append(('fan.speed', '%'))
+
+        if self._log_stats.temperature:
+            stat_keys.extend([('temperature.gpu', '°C'), ('temperature.memory', '°C')])
+
+        return stat_keys