+++ pytorch_lightning/trainer/evaluation_loop.py
@@ -124,31 +124,15 @@ In this second case, the options you pass to trainer will be used when running
 """
 
 from abc import ABC, abstractmethod
-from pprint import pprint
-from typing import Callable, List, Union
+from typing import Callable, List
 
 import torch
 from torch.utils.data import DataLoader
 
 from pytorch_lightning.core.lightning import LightningModule
-from pytorch_lightning.utilities import rank_zero_warn, flatten_dict, AMPType
-from pytorch_lightning.core.step_result import EvalResult, Result
+from pytorch_lightning.utilities import AMPType
 from pytorch_lightning.trainer.evaluate_loop import EvaluationLoop
-
-try:
-    import torch_xla.distributed.parallel_loader as xla_pl
-    import torch_xla.core.xla_model as xm
-except ImportError:
-    XLA_AVAILABLE = False
-else:
-    XLA_AVAILABLE = True
-
-try:
-    import horovod.torch as hvd
-except (ModuleNotFoundError, ImportError):
-    HOROVOD_AVAILABLE = False
-else:
-    HOROVOD_AVAILABLE = True
+from pytorch_lightning.trainer.logger_connector import LoggerConnector
 
 
 class TrainerEvaluationLoopMixin(ABC):
@@ -180,6 +164,7 @@ class TrainerEvaluationLoopMixin(ABC):
     verbose_test: bool
     running_sanity_check: bool
     amp_backend: AMPType
+    logger_connector: LoggerConnector
 
     # Callback system
     on_validation_batch_start: Callable
@@ -194,31 +179,7 @@ class TrainerEvaluationLoopMixin(ABC):
     evaluation_loop: EvaluationLoop
 
     @abstractmethod
-    def copy_trainer_model_properties(self, *args):
-        """Warning: this is just empty shell for code implemented in other class."""
-
-    @abstractmethod
     def get_model(self) -> LightningModule:
-        """Warning: this is just empty shell for code implemented in other class."""
-
-    @abstractmethod
-    def transfer_batch_to_gpu(self, *args):
-        """Warning: this is just empty shell for code implemented in other class."""
-
-    @abstractmethod
-    def add_progress_bar_metrics(self, *args):
-        """Warning: this is just empty shell for code implemented in other class."""
-
-    @abstractmethod
-    def log_metrics(self, *args, **kwargs):
-        """Warning: this is just empty shell for code implemented in other class."""
-
-    @abstractmethod
-    def reset_test_dataloader(self, *args):
-        """Warning: this is just empty shell for code implemented in other class."""
-
-    @abstractmethod
-    def reset_val_dataloader(self, *args):
         """Warning: this is just empty shell for code implemented in other class."""
 
     @abstractmethod
@@ -287,14 +248,11 @@ class TrainerEvaluationLoopMixin(ABC):
         eval_results = self.evaluation_loop.evaluation_epoch_end(num_dataloaders=len(dataloaders))
 
         # bookkeeping
-        self.evaluation_loop.log_epoch_metrics(eval_results)
+        eval_loop_results = self.evaluation_loop.log_epoch_metrics(eval_results, test_mode)
         self.evaluation_loop.predictions.to_disk()
 
         # hook
         self.evaluation_loop.on_evaluation_epoch_end()
-
-        # log the final eval loop metrics
-        eval_loop_results = self.__log_evaluation_epoch_metrics(eval_results, test_mode)
 
         # enable train mode again
         model.train()
@@ -304,51 +262,3 @@ class TrainerEvaluationLoopMixin(ABC):
         self.evaluation_loop.on_evaluation_end()
 
         return eval_loop_results, eval_results
-
-    def __log_evaluation_epoch_metrics(self, eval_results, test_mode):
-        if self.running_sanity_check:
-            return
-
-        eval_loop_results = []
-        if eval_results is not None and len(eval_results) > 0:
-
-            # in eval, the user may return something at every validation step without final reduction
-            if not isinstance(eval_results, list):
-                eval_results = [eval_results]
-
-            for result_idx, result in enumerate(eval_results):
-                if isinstance(result, EvalResult):
-                    prog_bar_metrics = result.epoch_pbar_metrics
-                    log_metrics = result.epoch_log_metrics
-                    callback_metrics = result.callback_metrics
-
-                    # in testing we don't need the callback metrics
-                    if test_mode:
-                        callback_metrics = {}
-                else:
-                    _, prog_bar_metrics, log_metrics, callback_metrics, _ = self.process_output(result)
-
-                # eval loop returns all metrics
-                dataloader_result_metrics = {**prog_bar_metrics, **log_metrics, **callback_metrics}
-
-                # add metrics to prog bar
-                self.add_progress_bar_metrics(prog_bar_metrics)
-
-                # log metrics
-                self.log_metrics(log_metrics, {})
-
-                # track metrics for callbacks
-                self.callback_metrics.update(callback_metrics)
-
-                if len(dataloader_result_metrics) > 0:
-                    eval_loop_results.append(dataloader_result_metrics)
-
-        # log results of test
-        if test_mode and self.is_global_zero and self.verbose_test:
-            print('-' * 80)
-            for result_idx, results in enumerate(eval_loop_results):
-                print(f'DATALOADER:{result_idx} TEST RESULTS')
-                pprint(results)
-                print('-' * 80)
-
-        return eval_loop_results