+++ tests/test_deprecated.py
@@ -6,7 +6,7 @@ import pytest
 import torch
 
 from pytorch_lightning import Trainer
-from pytorch_lightning.callbacks import LearningRateLogger
+from pytorch_lightning.callbacks import GpuUsageLogger, LearningRateLogger
 from tests.base import EvalModelTemplate
 
 
@@ -19,6 +19,12 @@ def _soft_unimport_module(str_module):
 def test_tbd_remove_in_v0_11_0_trainer():
     with pytest.deprecated_call(match='will be removed in v0.11.0'):
         lr_logger = LearningRateLogger()
+
+
+@pytest.mark.skipif(not torch.cuda.is_available(), reason="test requires GPU machine")
+def test_tbd_remove_in_v0_11_0_trainer_gpu():
+    with pytest.deprecated_call(match='will be removed in v0.11.0'):
+        gpu_usage = GpuUsageLogger()
 
 
 def test_tbd_remove_in_v0_10_0_trainer():
@@ -106,7 +112,7 @@ class ModelVer0_7(EvalModelTemplate):
 #     with pytest.deprecated_call(match='will be removed in v1.0. Use `test_epoch_end` instead'):
 #         trainer = Trainer(logger=False)
 #         trainer.test(model)
-#     assert trainer.callback_metrics == {'test_loss': torch.tensor(0.6)}
+#     assert trainer.logger_connector.callback_metrics == {'test_loss': torch.tensor(0.6)}
 #
 #     with pytest.deprecated_call(match='will be removed in v1.0. Use `validation_epoch_end` instead'):
 #         trainer = Trainer(logger=False)
@@ -119,7 +125,7 @@ class ModelVer0_7(EvalModelTemplate):
 #     with pytest.deprecated_call(match='will be removed in v1.0. Use `test_epoch_end` instead'):
 #         trainer = Trainer(logger=False)
 #         trainer.test(model)
-#     assert trainer.callback_metrics == {'test_loss': torch.tensor(0.7)}
+#     assert trainer.logger_connector.callback_metrics == {'test_loss': torch.tensor(0.7)}
 #
 #     with pytest.deprecated_call(match='will be removed in v1.0. Use `validation_epoch_end` instead'):
 #         trainer = Trainer(logger=False)